Added BrowserServer support.
QPA support for building QT. BrowserServer compilation tweaks as
it needs to be compiled with -rpath not -rpath-link otherwise
the runtime linker fails to find libpbnjson_cpp.so. Removed opengl
from the build as that was causing problems with XCB.

Startup script now runs the BrowserServer before LunaSysMgr

Added new dependencies to README.md

Open-webOS-DCO-1.0-Signed-off-by: Mikko Levonmaa <mikko.levonmaa@palm.com>

+++ build-webos-desktop.sh
@@ -17,7 +17,6 @@
 #
 # LICENSE@@@
 
-
 if [ "$1" = "clean" ] ; then
   export SKIPSTUFF=0
   set -e
@@ -250,9 +249,11 @@ function build_qt4
       cd $BASE/qt-build-desktop
       if [ ! -e ../qt4/palm-desktop-configure.orig ] ; then
         cp -f ../qt4/palm-desktop-configure ../qt4/palm-desktop-configure.orig
-        sed -i 's/-opensource/-opensource -fast -qconfig palm -no-dbus/' ../qt4/palm-desktop-configure
+        sed -i 's/-opensource/-opensource -qpa -fast -qconfig palm -no-dbus/' ../qt4/palm-desktop-configure
         sed -i 's/libs tools/libs/' ../qt4/palm-desktop-configure
       fi
+      # This export will be picked up by plugins/platforms/platforms.pro and xcb.pro
+      export WEBOS_CONFIG="webos desktop"
       ../qt4/palm-desktop-configure
     fi
     cd $BASE/qt-build-desktop
@@ -261,8 +262,10 @@ function build_qt4
 
     # Make alias to moc for BrowserServer build
     # (Could also fix w/sed in BrowserServer build for Makefile.Ubuntu)
-    cd ${LUNA_STAGING}/bin
-    ln -s moc-palm moc
+    if [ ! -e ${LUNA_STAGING}/bin/moc ]; then
+        cd ${LUNA_STAGING}/bin
+        ln -s moc-palm moc
+    fi
 }
 
 ################################
@@ -852,8 +855,9 @@ function build_BrowserServer
     export STAGING_DIR=${LUNA_STAGING}
     export STAGING_INCDIR="${LUNA_STAGING}/include"
     export STAGING_LIBDIR="${LUNA_STAGING}/lib"
-    # link fails without -rpath-link to help liblunaservice find libcjson
-    export LDFLAGS="-Wl,-rpath-link $LUNA_STAGING/lib"
+    # link fails without -rpath to help liblunaservice find libcjson
+    # and with rpath-link it fails ar runtime
+    export LDFLAGS="-Wl,-rpath $LUNA_STAGING/lib"
     make $JOBS -e PREFIX=$LUNA_STAGING -f Makefile.Ubuntu all BUILD_TYPE=release
 
     # stage files
